requirements.txt
beir>=2.0.0
numpy>=1.23
rank-bm25>=0.2.2
tqdm>=4.66.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
torch>=2.3.0
